backend/requirements.txt
Flask>=2.3,<3.1
Werkzeug>=2.3,<3.1
flask-cors>=4.0,<5
python-dotenv>=1.0,<2.0

PyJWT>=2.8,<3
bcrypt>=4.1,<5

pymongo>=4.6,<5
dnspython>=2.6,<3

# ML / text processing (reused from your project)
# numpy 2.x wheels ship for Python 3.13; numpy 1.x often tries to build from source on 3.13
scikit-learn==1.7.2
joblib>=1.3,<2.0
numpy>=2.0,<3.0
pandas>=2.1,<3.0
nltk>=3.8,<4.0
textblob>=0.17,<0.19
spacy>=3.7,<3.9
wikipedia-api>=0.8,<1.0
requests>=2.31,<3.0
beautifulsoup4>=4.12,<5.0
feedparser>=6.0,<7.0
shap>=0.44,<0.50

# Optional fact-checker visuals (kept to avoid runtime import issues)
matplotlib>=3.6,<4.0
seaborn>=0.12,<0.14
plotly>=5.15,<6.0
wordcloud>=1.9,<2.0

